:sparkles: binding textures

+++ TextureManager.cpp
@@ -45,7 +45,7 @@ void TextureManager::CreateSRVDescriptorHeap(ID3D12Device* device)
 	LOG_FUNCTION_ENTRY();
 	D3D12_DESCRIPTOR_HEAP_DESC srvHeapDesc = {};
 	srvHeapDesc.Type = D3D12_DESCRIPTOR_HEAP_TYPE_CBV_SRV_UAV;
-	srvHeapDesc.NumDescriptors = m_textureMap.size();
+	srvHeapDesc.NumDescriptors = static_cast<UINT>(m_textureMap.size());
 	srvHeapDesc.Flags = D3D12_DESCRIPTOR_HEAP_FLAG_SHADER_VISIBLE;
 	srvHeapDesc.NodeMask = 0;
 	THROW_IF_FAILED(device->CreateDescriptorHeap(&srvHeapDesc, IID_PPV_ARGS(m_srvHeap.GetAddressOf())))
